editor migration cleanup
