+++ app.ts
@@ -78,13 +78,6 @@ namespace microcode {
             const hostedProgram = this.uiHost.currentEditorProgram()
             if (hostedProgram) {
                 runProgramIfStopped(hostedProgram)
-                return
-            }
-            const topIndex = this.sceneManager.scenes.length - 1
-            const topScene = this.sceneManager.scenes[topIndex]
-            if (topScene instanceof Editor) {
-                const editor: Editor = topScene
-                editor.runProgram()
             }
         }
     }
