+++ pxt.json
@@ -19,11 +19,10 @@
         "tiles.ts",
         "app.ts",
         "apphost.ts",
+        "storage.ts",
         "hosted/home.ts",
         "hosted/gallery.ts",
         "hosted/editor.ts",
-        "editor.ts",
-        "ruleeditor.ts",
         "version.ts",
         "jacs.ts",
         "fieldeditors.ts",
