+++ storage.ts
@@ -1,0 +1,8 @@
+namespace microcode {
+    /**
+     * Save-slot ids shown by Home load and editor save pickers.
+     */
+    export function diskSlots() {
+        return ["disk1", "disk2", "disk3"]
+    }
+}
